+++ workers/unity/Packages/io.improbable.gdk.tools/.CodeGenTemplate/CodeGenerationLib/CodeWriter/Interfaces/ICodeBlock.cs
@@ -1,0 +1,10 @@
+namespace Improbable.Gdk.CodeGeneration.CodeWriter
+{
+    /// <summary>
+    /// A unit of code that can be returned as an indented string.
+    /// </summary>
+    public interface ICodeBlock
+    {
+        string Format(int indentLevel = 0);
+    }
+}