+++ workers/unity/Packages/io.improbable.gdk.tools/.CodeGenTemplate/CodeGenerationLib/CodeWriter/Scopes/LoopBlock.cs
@@ -1,0 +1,25 @@
+using System;
+
+namespace Improbable.Gdk.CodeGeneration.CodeWriter.Scopes
+{
+    /// <summary>
+    /// Scope with mandatory declaration for defining loops.
+    /// </summary>
+    public class LoopBlock : ScopeBody
+    {
+        internal LoopBlock(string declaration, Action<LoopBlock> populate) : base(declaration)
+        {
+            populate(this);
+        }
+
+        public void Continue()
+        {
+            Line("continue;");
+        }
+
+        public void Break()
+        {
+            Line("break;");
+        }
+    }
+}