+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Internal
+
+- Implemented a new CodeWriter in the code generator which provides a fluent interface for generating C# code. [#1237](https://github.com/spatialos/gdk-for-unity/pull/1237)
+    - The `CodeGenerationLib` has been migrated to the new `CodeWriter`.
+
 ## `0.3.2` - 2019-12-23
 
 ### Added